refactor: implement allReferenceStore to pages (project, projectGroup, user)
