+++ apps/web/src/services/cost-explorer/components/BudgetDetailInfo.vue
@@ -13,9 +13,9 @@ import { ACTION_ICON } from '@spaceone/design-system/src/inputs/link/type';
 import type { BudgetModel } from '@/schema/cost-analysis/budget/model';
 import { store } from '@/store';
 
-import type { ProjectReferenceMap } from '@/store/modules/reference/project/type';
 import type { ProviderReferenceMap } from '@/store/modules/reference/provider/type';
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
+import type { ProjectReferenceMap } from '@/store/reference/project-reference-store';
 import type { WorkspaceReferenceMap } from '@/store/reference/workspace-reference-store';
 
 import { referenceRouter } from '@/lib/reference/referenceRouter';
