+++ apps/web/src/services/home-dashboard/components/TrustedAdvisor.vue
@@ -119,14 +119,15 @@ import { i18n } from '@/translations';
 
 import type { FavoriteItem } from '@/store/modules/favorite/type';
 import { FAVORITE_TYPE } from '@/store/modules/favorite/type';
-import type { ProjectReferenceMap } from '@/store/modules/reference/project/type';
 import type { ProviderReferenceMap } from '@/store/modules/reference/provider/type';
+import type { ProjectReferenceMap } from '@/store/reference/project-reference-store';
 
 import WidgetLayout from '@/common/components/layouts/WidgetLayout.vue';
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { ASSET_INVENTORY_ROUTE } from '@/services/asset-inventory/routes/route-constant';
 import TrustedAdvisorOverall from '@/services/home-dashboard/components/TrustedAdvisorOverall.vue';
+
 
 enum STATUS {
     error = 'error',
@@ -284,7 +285,6 @@ export default {
             await Promise.allSettled([
                 store.dispatch('favorite/load', FAVORITE_TYPE.PROJECT),
                 // LOAD REFERENCE STORE
-                store.dispatch('reference/project/load'),
                 store.dispatch('reference/provider/load'),
             ]);
         };
