+++ apps/web/src/services/project/components/ProjectMainProjectGroupFormModal.vue
@@ -15,8 +15,6 @@ import type { ProjectGroupUpdateParameters } from '@/schema/identity/project-gro
 import type { ProjectGroupModel } from '@/schema/identity/project-group/model';
 import { i18n } from '@/translations';
 
-import { useProjectGroupReferenceStore } from '@/store/reference/project-group-reference-store';
-
 import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
@@ -25,7 +23,6 @@ import { useFormValidator } from '@/common/composables/form-validator';
 import { useProjectPageStore } from '@/services/project/stores/project-page-store';
 
 
-const projectGroupStore = useProjectGroupReferenceStore();
 const projectPageStore = useProjectPageStore();
 const projectPageGetters = projectPageStore.getters;
 const projectPageState = projectPageStore.state;
@@ -72,7 +69,6 @@ const getProjectGroupNames = async () => {
 const createProjectGroup = async (params: ProjectGroupCreateParameters) => {
     try {
         await projectPageStore.createProjectGroup(params);
-        await projectGroupStore.load({ force: true });
         showSuccessMessage(i18n.t('PROJECT.LANDING.ALT_S_CREATE_PROJECT_GROUP'), '');
     } catch (e) {
         ErrorHandler.handleRequestError(e, i18n.t('PROJECT.LANDING.ALT_E_CREATE_PROJECT_GROUP'));
@@ -81,7 +77,6 @@ const createProjectGroup = async (params: ProjectGroupCreateParameters) => {
 const updateProjectGroup = async (params: Partial<ProjectGroupUpdateParameters>) => {
     try {
         await projectPageStore.updateProjectGroup(params);
-        await projectGroupStore.load({ force: true });
         showSuccessMessage(i18n.t('PROJECT.LANDING.ALT_S_UPDATE_PROJECT_GROUP'), '');
     } catch (e) {
         ErrorHandler.handleRequestError(e, i18n.t('PROJECT.LANDING.ALT_E_UPDATE_PROJECT_GROUP'));
